Access Permissions, Security Layers added
basic security layers added anyone cannot use the application from the URL directly. if you want to create a new post then you have to be a user of this application. and if you are not an user then you can only see the posts but cannot edit and delete them.

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -47,8 +57,8 @@ class PostsController extends Controller
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
         $post->save();
-
-        return redirect('/posts')->with('success', 'Post created');
+        
+        return redirect('/dashboard')->with('success', 'Post created');
     }
 
     /**
@@ -72,6 +82,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Checking for current User
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');    
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -107,8 +122,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        // Checking for current User
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');    
+        }
         $post->delete();
 
-        return redirect('/posts')->with('success', 'Post Deleted');
+        return redirect('/dashboard')->with('success', 'Post Deleted');
     }
 }
